requirements.txt
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
# Нативный диалог «файл или папка» на macOS (опционально)
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
